livekit-voice-agent/pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "livekit-agents[google,silero,turn-detector]~=1.2",
    "livekit-plugins-deepgram>=1.3.6",
    "livekit-plugins-elevenlabs>=1.3.6",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-openai>=1.3.6",
    "python-dotenv>=1.2.1",
]
